01_09_fastapi_backend/myfastapi-start/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
psycopg2-binary==2.9.5
pydantic
loguru